moved db sub-command related code in the db dir and move Config into its own file. Updated the result report to give back a little more information
