+++ src/ports/storageservice.rs
@@ -11,7 +11,4 @@ pub trait StorageService {
     /// stores the data in the store that implements this trait, a table in relational databases but
     /// returns the number of records stored successfully or any error(s) the method encounters
     fn store_data(&self, column_defs: Vec<ColumnDef>, data: Vec<StringRecord>, insert_stmt: String) -> Result<usize, Error>;
-    /// validates the number of records that existed in the CSV file were added to the store
-    /// returns the true if the total_lines is equal to the number of records in the store
-    fn validate(&self,name: String, total_lines: usize) -> Result<bool, Error>;
 }
